(BKR-763) Supply only necessary answers on upgrade
Previous to this commit, answer files were generated in full during
upgrade scenarios; this does not match the user workflow or what is
documented for upgrades, so we should only supply answers needed. The
answers `create` method now checks the `:type` in options and creates a
separate branch of `Answers` named `Upgrade`.

This change only affects PE 3.8 and higher; all lower versions will
continue to produce an entire answers file.