(PE-14996) Add DB user/name to default hiera configs
Prior to this commit we were not specifying the old default
DB user/names in the hiera configs, which could result in tests
that depend on those names to fail. Additionally, it is a good
idea to verify that specifying a non-default database name still
works.
This commit specifies puppetdb, classifier, activity, rbac and
orchestrator database names by default for the 2016.2 hiera
config/answer files. console/console_auth are omitted because
those databases are deprecated in 2016.2.

Also fixed a typo where `q_database_name` was specified in the
defaults hash instead of `q_classifier_database_name` and fixed
the only reference to that (from 3.4).