(PE-14996) Further refactor default config generation
Prior to this commit there was a large amount of duplicate code
used for getting the `answer_for` for each key.

This commit refactors that duplicate code out into a method on
the Answers class that takes an array of desired defaults and
returns an array with either the default from the Answers class
or the user-provider override.

Additionally, instead of them hiera answers, we should be calling
them hiera config in order to be consistent with how we are documenting
them and what they really are; persistent configuration that will be
saved on the system.