(PE-14996) Use case for determining answer type
Prior to this commit we were using a series of if statements
when checking if we support the answer `type` provided.
This commit changes those checks to use a case statement in order
to clean up the logic. Additionally, more exceptions will be raised
if methods are called with an unknown `type`.