(PE-14996) Use :format in options instead of @type as param
Prior to this commit I was adding a type param to the init/create
methods of `Answers` in order to specify which answers file type
to generate. Using the word type, which is already used to specify
`:install` vs `:upgrade`, and creating a whole new param were
somewhat counter-intuitive and confusing.

This commit removes the type param, and instead reads the `:format`
key from the `options` hash upon initialization, and if it cannot find
a value at that key, it will default to `:bash`. I went back to using
`:format` in order to reduce the chance of us later being confused
by the `:type` option and decided to use the options hash because that
is how we have typically interacted with the `Answers` library in the
past. This should hopefully be a more intuitive and less error prone
way of interacting with the library. Fingers crossed.