(PE-14996) Provide a format agnostic method for configuration string
With the changes in installer configuration format, Version201620 now
has an answer_hiera method.  However the component using beaker-answers
shouldn't have to care about which output method it calls to get a
configuration.  Ultimately it should just expect the correct
configuration for a given PE version.  To assist with this, this commit
adds an installer_configuration_string(host) method which will return
either an answer file string or a hiera pe.conf string depending on the
:format setting.

Because no version prior to 2016.2.0 can work with a hiera pe.conf, I've
moved the implementation into Version201620 and left an erroring stub in
Answers.