(PE-14996) Fix test for legacy answers in hiera_config to handle symbols
Beaker configs in this format:

:answers:
  :q_puppet_enterpriseconsole_auth_password: puppetlabs
  :q_puppet_enterpriseconsole_auth_user_email: admin@puppetlabs.net
CONFIG:
  log_level: debug
HOSTS:
  master.vm:
    platform: el-7-x86_64
...

where legacy answers are provided to be slurped into Beaker options and
passed on to BeakerAnswers, were failing when BeakerAnswers was
attempting to produce a 2016.2.x hiera config for meep.  This was just
because the test was assuming strings for keys instead of symbols.  The
patch converts to string before testing and adds a spec.