(PE-14996) Treat q_puppetdb_database_password default same as others
181305f6ba6116d05c8163eb2a18243df30098d0 changed beaker-answers so that
the puppetdb default was referenced by its answer
q_puppetdb_database_password, rather than the key q_puppetdb_password,
which is not used in the legacy answer file.  Due to how answer_for()
works, where an existing value looked up from DEFAULTS prevents the
passed default override from being used, this caused the generated
answer for puppetdb to added without quotes. To fix this, quotes were
added to the default.  This worked, but it ran counter to how the later
database answers for classifier, rbac, activity and orchestrator
database passwords were handled.  These had defaults without quotes, and
then explicitly quoted the derived answers generated for the
answers_hash.

When we began using answers for generating a hiera config for meep, this
caused problems with the puppetdb password because the quotes in the
puppetdb default ended up in the json value, causing an error in
Postgres.

To fix this, and to unify q_puppetdb_database_password's handling with
that of the other db passwords, this commit removes the quoting in the
defaults, and adds it to the generation of the final
q_puppetdb_database_password answer.  It also modified a spec that was
testing the generation of this answer to take into account that the
quotes are added by the answer generation process.