(PE-14996) Change upgrade answers spec to be literal strings
Even though we reverted the DEFAULT_ANSWERS change to no longer have
quotes within the answer values, the spec tests were still passing, even
though they would generate erroneous answers for the
puppetdb_database_password. This change makes it so that the tests are
testing for literal strings, as opposed to just ensuring the calculation
for the answer happens.