(PE-15259) Only include legacy database user defaults if prompted
If beaker-pe initializes BeakerAnswers with an
:include_legacy_database_defaults, then add custom database users from
BeakerAnswers::DEFAULT_ANSWERS.  This is only necessary if we are
upgrading from a pre-meep Beaker install of PE where BeakerAnswers
changed the default user names.

Setting the old legacy database password defaults is never required.

To override any of these settings, they should be added directly to the
:answers hash in the host.cfg file.