(PE-14996) Ensure flatten_keys_to_joined_string generates strings
Prior to this commit I didn't ensure that keys being given in hashes
to `flatten_keys_to_joined_string` were being converted to strings.
This would result in keys that were passed in as symbols acting as a
duplicate key to the string equivalent, which were then written out
by the json/hocon parsers because it seems they do not have a uniqueness
constraint on keys in hashes.
This commit changes `flatten_keys_to_joined_string` to always convert
the keys to strings, which should prevent this sort of duplicate key
issue.