(PE-16028) Update required answers for 2016.2
Previous to this commit, answer generation for 2016.2 would generate a
list of parameters on the puppet_enterprise for both mono and split. Due
to changes introduced in Puppet 4.4, the puppet_enterprsie module no
longer needs all host parameters in a monolithic install case.
This commit updates the generation to match what the module has changed
to use. For a monolithic install, the only required answers are now the
console admin password, and puppet_master_host. for split installs with
PE postgres, just master, console and puppetdb host.