(PE-16065) Do not fall back to ip address when setting puppetdb hostname
Before this commit the puppetdb hostname answer would fallback to being
an ip address which would cause the installer to do incorrect things as
we depend on those answers to be hostnames for certificate generation
and configuration bootstrap purposes.

This commit changes the q_puppetdb_hostname answer to be the hostname of
the puppetdb node instead of its ip address which is returned by the
reachable_name method on the host object.