(PE-19203,PE-17621) Generate a meep 2.0 format based on version flag
Previously there was a hard version break set on PE 2017.1 which would
produce a 2.0 format pe.conf, but since we've had to back out meep
classification in Flanders, this needs to change.

Added a PeConf object to generate 1.0 and 2.0 format files based on a
meep_schema_version flag. It defaults to the original 1.0 format. It
will be up to beaker-pe to submit this flag in the beaker options based
on its use_meep_for_classification?() test.

There is one change to the 1.0 hiera_host_config implementation in that
it no longer uses answer_for to set the master, db, console settings. I
don't think this matters because there are no defaults for these
settings, and beaker answer overrides are merged afterwards in
generate_hiera_config() regardless.