(maint) Do not flatten answers
When generating meep 2.0 answers.

Changes in version20162.rb took an answers structure like:

:answers
  puppet_enterprise:
    profile:
      "master::foo": bar

and produce a pe.conf with

"puppet_enterprise::profile::master::foo": "bar"

This might be a convenience for some kinds of parameter listing, but
breaks your ability to submit actually structured data such as:

:answers
  feature_flags:
    pe_modules_next: true

Where feature_flags is supposed to be a hash, and not a parameter.

This pr takes the approach of just fixing new behavior in Glisan, when
using Meep 2.0 format (pe-modules-next), but it might be better to
change the behavior altogether, although this is a breaking change for
anyone relying on the previous 'helpful' collapsing of answer elements.