(PE-19948) Fix pe_repo platform class for windows
Before this commit the pe_repo platform class that would be generated
for all windows platforms would be incorrect as there is only a single
windows puppet agent. When this method would generate the list of
platform classes windows would be incorrect.

This commit ensures that when windows is a platform on a host we return
one of the two possible windows pe_repo platform classes.