(PE-22724) Add PE managed postgres support
This PR adds support for PE managed postgres support.
It detects for a new role: pe_postgres.
If that role is found, that node will be set as the puppetdb PE node.
Then there will be a check for how to configure the postgres node.
Either way SSL or with password, depending on options supplied.
The meep 2.0 settings are a bit of work in progress, future work
will probably be required to get them just right.