(PE-22724) Add puppetdb startout timer for external postgres
This setting was left out in my previous commit. Without this setting
on a monolithic install with PE managed external postgres, the
puppet database takes too long (25 minutes+) to time out. This causes
failures in Jenkins.
This answer that is added in will make the timeout only a few minutes
so hopefuly Jenkins wont abort the job.