(maint) Constrain overly-permissive version checking
A patch was made so that beaker-answers would support any version past 2017.3 without requiring changes to this library each time a new release was created. However, the regex used to detect versions up to 9999.9.99 caused it to also match pre-2017.3 versions, which could cause unsupported answers to be used. This constrains the version checking to limit the maximum version to 2099.9.99. We'll probably all be dead by then, along with this code.